+++ cognite/client/data_classes/events.py
@@ -4,6 +4,29 @@ from cognite.client.data_classes._base import *
 from cognite.client.data_classes.shared import TimestampRange
 
 
+# GenPropertyClass: EndTimeFilter
+class EndTimeFilter(dict):
+    """Either range between two timestamps or isNull filter condition.
+
+    Args:
+        max (int): The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds.
+        min (int): The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds.
+        is_null (bool): Set to true if you want to search for data with field value not set, false to search for cases where some value is present.
+    """
+
+    def __init__(self, max: int = None, min: int = None, is_null: bool = None, **kwargs):
+        self.max = max
+        self.min = min
+        self.is_null = is_null
+        self.update(kwargs)
+
+    max = CognitePropertyClassUtil.declare_property("max")
+    min = CognitePropertyClassUtil.declare_property("min")
+    is_null = CognitePropertyClassUtil.declare_property("isNull")
+
+    # GenStop
+
+
 # GenClass: Event
 class Event(CogniteResource):
     """An event represents something that happened at a given interval in time, e.g a failure, a work order etc.
@@ -16,7 +39,7 @@ class Event(CogniteResource):
         type (str): Type of the event, e.g 'failure'.
         subtype (str): Subtype of the event, e.g 'electrical'.
         description (str): Textual description of the event.
-        metadata (Dict[str, str]): Custom, application specific metadata. String key -> String value. Limits: Maximum length of key is 32 bytes, value 512 bytes, up to 16 key-value pairs.
+        metadata (Dict[str, str]): Custom, application specific metadata. String key -> String value. Limits: Maximum length of key is 128 bytes, value 128000 bytes, up to 256 key-value pairs, of total size at most 200000.
         asset_ids (List[int]): Asset IDs of equipment that this event relates to.
         source (str): The source of this event.
         id (int): A server-generated ID for the object.
@@ -66,8 +89,9 @@ class EventFilter(CogniteFilter):
 
     Args:
         start_time (Union[Dict[str, Any], TimestampRange]): Range between two timestamps.
-        end_time (Union[Dict[str, Any], TimestampRange]): Range between two timestamps.
-        metadata (Dict[str, str]): Custom, application specific metadata. String key -> String value. Limits: Maximum length of key is 32 bytes, value 512 bytes, up to 16 key-value pairs.
+        end_time (Union[Dict[str, Any], EndTimeFilter]): Either range between two timestamps or isNull filter condition.
+        active_at_time (Union[Dict[str, Any], TimestampRange]): Event is considered active from its startTime to endTime inclusive. If startTime is null, event is never active. If endTime is null, event is active from startTime onwards. activeAtTime filter will match all events that are active at some point from min to max, from min, or to max, depending on which of min and max parameters are specified.
+        metadata (Dict[str, str]): Custom, application specific metadata. String key -> String value. Limits: Maximum length of key is 128 bytes, value 128000 bytes, up to 256 key-value pairs, of total size at most 200000.
         asset_ids (List[int]): Asset IDs of equipment that this event relates to.
         asset_external_ids (List[str]): Asset External IDs of equipment that this event relates to.
         root_asset_ids (List[Dict[str, Any]]): This parameter is deprecated. Use assetSubtreeIds instead. Only include events that have a related asset in a tree rooted at any of these root assetIds.
@@ -85,7 +109,8 @@ class EventFilter(CogniteFilter):
     def __init__(
         self,
         start_time: Union[Dict[str, Any], TimestampRange] = None,
-        end_time: Union[Dict[str, Any], TimestampRange] = None,
+        end_time: Union[Dict[str, Any], EndTimeFilter] = None,
+        active_at_time: Union[Dict[str, Any], TimestampRange] = None,
         metadata: Dict[str, str] = None,
         asset_ids: List[int] = None,
         asset_external_ids: List[str] = None,
@@ -102,6 +127,7 @@ class EventFilter(CogniteFilter):
     ):
         self.start_time = start_time
         self.end_time = end_time
+        self.active_at_time = active_at_time
         self.metadata = metadata
         self.asset_ids = asset_ids
         self.asset_external_ids = asset_external_ids
@@ -123,7 +149,9 @@ class EventFilter(CogniteFilter):
             if instance.start_time is not None:
                 instance.start_time = TimestampRange(**instance.start_time)
             if instance.end_time is not None:
-                instance.end_time = TimestampRange(**instance.end_time)
+                instance.end_time = EndTimeFilter(**instance.end_time)
+            if instance.active_at_time is not None:
+                instance.active_at_time = TimestampRange(**instance.active_at_time)
             if instance.created_time is not None:
                 instance.created_time = TimestampRange(**instance.created_time)
             if instance.last_updated_time is not None:
@@ -208,6 +236,23 @@ class _ListEventUpdate(CogniteListUpdate):
 
     def remove(self, value: List) -> EventUpdate:
         return self._remove(value)
+
+    # GenStop
+
+
+# GenPropertyClass: EventAggregate
+class EventAggregate(dict):
+    """Aggregation group of events
+
+    Args:
+        count (int): Size of the aggregation group
+    """
+
+    def __init__(self, count: int = None, **kwargs):
+        self.count = count
+        self.update(kwargs)
+
+    count = CognitePropertyClassUtil.declare_property("count")
 
     # GenStop